Enable a bunch of clang-tidy checks
New year, new set of clang-tidy checks we can enable.  This clang-tidy
list was brought over from bmcweb.  Most of the fixes are relatively
trivial.

This probably should be broke down more.  As-is, it's kind of a big
change.

Tested: TBD

+++ src/PSUSensorMain.cpp
@@ -818,7 +818,7 @@ static void createSensorsCallback(
             if (!customizedScale)
             {
                 // Preserve existing usage of hardcoded labelMatch table below
-                factor = std::pow(10.0, factor);
+                factor = static_cast<unsigned int>(std::pow(10.0, factor));
 
                 /* Change first char of substring to uppercase */
                 char firstChar =
